Clean up integrationt test
